add error handling

+++ pydynamo/storage/inmemorystore.py
@@ -1,7 +1,8 @@
 from .iterator import Iterator
 from .binarysearchtree import BinarySearchTree
 from .store import Store
-
+from .error import StorageException
+from .error import ErrorType
 
 class InMemoryStore(Store):
     def __init__(self, store_name):
@@ -16,6 +17,7 @@ class InMemoryStore(Store):
 
     def get(self, key: str) -> str:
         if not self.database.contain(key):
+            raise StorageException(ErrorType.NOT_FOUND, "This key cannot be found in store.")
             return None
         else:
             return self.database.get(key)
@@ -23,5 +25,7 @@ class InMemoryStore(Store):
     def iterator(self) -> Iterator:
         return Iterator(self.database)
 
-    def remove(self, key:str) -> None:
+    def remove(self, key: str) -> None:
+        if not self.database.contain(key):
+            raise StorageException(ErrorType.NOT_FOUND, "This key cannot be found in store.")
         self.database.remove_node(key)
